Fuck animations, but the logic works!
I think I broke some of the fundamental logic when reorganizing, and the async/await + Awaitable quirks were NOT HELPING debug. So I banished them to the wastelands...
I will return, but I think the goal of "pretty from day one" is gonna have to be dropped. It's not worth the hassle. Too much time lost.